+++ pyxem/signals/diffraction_vectors.py
@@ -16,7 +16,6 @@
 # You should have received a copy of the GNU General Public License
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
 
-import itertools
 from functools import cached_property
 from warnings import warn
 
@@ -47,6 +46,13 @@ from pyxem.utils.vector_utils import (
 )
 
 from pyxem.utils._slicers import Slicer
+
+from pyxem.utils._subpixel_utils import (
+    _conventional_xc_map,
+    _center_of_mass_map,
+    _get_simulated_disc,
+)
+
 from pyxem.utils._deprecated import deprecated
 
 """
@@ -183,10 +189,10 @@ class DiffractionVectors(BaseSignal):
         """
         if center is None and peaks.metadata.has_item("Peaks.signal_axes"):
             center = [
-                -ax.offset / ax.scale for ax in peaks.metadata.Peaks.signal_axes[::-1]
+                ax.offset / ax.scale for ax in peaks.metadata.Peaks.signal_axes[::-1]
             ]
         elif center is not None:
-            pass  # center is already set
+            center = -np.array(center)
         else:
             raise ValueError(
                 "A center and calibration must be provided unless the"
@@ -243,7 +249,7 @@ class DiffractionVectors(BaseSignal):
             units = list(units) + ["a.u."]
 
         vectors = peaks.map(
-            lambda x, cen, cal: (x - cen) * cal,
+            lambda x, cen, cal: (x + cen) * cal,
             cal=calibration,
             cen=center,
             inplace=False,
@@ -260,21 +266,112 @@ class DiffractionVectors(BaseSignal):
         vectors.column_names = column_names
         return vectors
 
+    def subpixel_refine(
+        self,
+        signal,
+        method="center-of-mass",
+        disk_r=None,
+        upsample_factor=2,
+        square_size=10,
+        **kwargs,
+    ):
+        """
+        Refine the positions of the diffraction vectors using subpixel
+        interpolation.
+
+        Parameters
+        ----------
+        signal : Signal
+            The signal which will be used to refine the diffraction vectors.
+        method : str
+            The method used to refine the diffraction vectors. Currently
+            supported methods are 'cross-correlation' and
+            and "center-of-mass".
+        disk_r : int or None
+            The radius of the disk used for the cross-correlation method in pixels.
+        upsample_factor : int
+            The upsample factor used for the cross-correlation method.
+        square_size : int
+            The size of the square used for both the center-of-mass and cross-correlation methods.
+        kwargs : dict
+            Additional keyword arguments to be passed to the map method.
+
+        Returns
+        -------
+        refined_vectors : DiffractionVectors
+            The refined vectors.
+        """
+        method_dict = {
+            "cross-correlation": _conventional_xc_map,
+            "center-of-mass": _center_of_mass_map,
+        }
+        if method not in method_dict:
+            raise ValueError(
+                f"The method parameter must be one of {list(method_dict.keys())}"
+            )
+        if method == "cross-correlation":
+            kwargs["upsample_factor"] = upsample_factor
+            kwargs["kernel"] = _get_simulated_disc(square_size, disk_r)
+        kwargs["square_size"] = square_size
+
+        signal_axes = signal.axes_manager.signal_axes
+        offsets = [ax.offset for ax in signal_axes]
+        scales = [ax.scale for ax in signal_axes]
+
+        funct = method_dict[method]
+        pixels = self.get_pixel_vectors(
+            offsets=offsets,
+            scales=scales,
+            shape=signal.axes_manager._signal_shape_in_array,
+            square_size=square_size,
+        )
+        refined_vectors = signal.map(
+            funct,
+            vectors=pixels,
+            inplace=False,
+            ragged=True,
+            offsets=offsets,
+            scales=scales,
+            **kwargs,
+        )
+        refined_vectors.set_signal_type("diffraction_vectors")
+        refined_vectors._set_up_vector(
+            scales=self.scales, column_names=self.column_names
+        )
+        return refined_vectors
+
     @property
     def pixel_vectors(self):
+        return self.get_pixel_vectors()
+
+    def get_pixel_vectors(
+        self, offsets=None, scales=None, square_size=None, shape=None
+    ):
         """Returns the diffraction vectors in pixel coordinates."""
-        if self.scales is None or self.center is None:
-            raise ValueError(
-                "The pixel vectors cannot be calculated without a calibration."
-            )
+        if offsets is None:
+            offsets = self.offsets
+        if scales is None:
+            scales = self.scales
+
+        def get_pixels(x, off, scale, square_size=None, shape=None):
+            pixels = np.round((x - off) / scale).astype(int)
+            if square_size is not None and shape is not None:
+                pixels = pixels[np.all(pixels > (square_size / 2) + 1, axis=1)]
+                pixels = pixels[
+                    np.all(np.array(shape) - pixels > (square_size / 2) + 1, axis=1)
+                ]
+            return pixels
+
         pixels = self.map(
-            lambda x, cen, cal: x / cal + self.center,
-            cal=self.scales,
-            cen=self.center,
+            get_pixels,
+            off=offsets,
+            scale=scales,
+            square_size=square_size,
+            shape=shape,
             inplace=False,
             ragged=True,
         )
-        return pixels.data
+        return pixels
 
     @property
     def _is_object_dtype(self):